Emoji candidate filtering based on introduction version
This CL adds complex checks for each candidate in segments, in EnvironmentalFilterRewriter. Performance impact is therefore, inevitable.

By profiling with session_handler_benchmark_test, the occupation for the execution of EnvironmentalFilterRewriter::Rewrite turns out to be about 0.15% in total execution time. Previously, the occupation was about 0.01~0.03%. However, 0.15% is only 110ms in 70s execution time. The impact can be said to be not so bad.

PiperOrigin-RevId: 472646261