Changed the prefix penalty calculation so that the prefix of a single entry will not be promoted.
If we use the target prefix candidate as the history to run immutable converter to calculate the penalty, unexpected candidates can be promoted with low cost.
For example, when we have a query "おつかれ", a dictionary entry "お疲れ", and prefix candidate "お", if we used "お" as the history to run converter, "疲れ" can be promoted as the rest of the entry "お疲れ" with the low cost.
In this case, the prefix candidate "お" can be promoted over "お疲れ", since the penalty cost for "お"　is very low.

PiperOrigin-RevId: 473948629