Move Util::IsAndroidPuaEmoji to user_history_predictor.cc
Mozc does not support Android PUA Emoji now. Because it is possible that they are still in user history, removing the check for Android PUA Emoji is not feasible.
However, having the function in Util is not nice;
1. It's only used from user_history_predictor.cc
2. Having it in Util is noisy.

In order to keep codes clean, in this CL, Util::IsAndroidPuaEmoji is moved to user_history_predictor's internal implementation. Now it is not visible outside.

PiperOrigin-RevId: 474710317