Explicitly cast individual un-processable wide characters to integer types before logging them.
In C++20, the underlying std::ostream's operator<< overloads accepting wide character types have been deleted, so the existing code no longer compiles.

This was done to avoid the many accidental cases of logging characters formatted as numbers and strings formatted as pointers, but these cases are actually intentionally logging characters as code points that cannot be interpreted as a character.

PiperOrigin-RevId: 500854154