Explicitly stream individual wide characters formatted as numbers.
In C++17, the underlying std::ostream can format only standard-width chars as characters, formatting wider characters of type char16_t or char32_t as numbers. In C++20, the overloads accepting these wide characters are explicitly deleted to avoid using such formatting unintentionally. Thus, this code no longer compiles.

Casting to integer types produces the same behavior in both versions as was produced in C++17 without the cast.

In this case, format as hex for easier human debugging since the test cases provide the streamed values in hex. Happy to change that if domain knowledge suggests otherwise is better.

This change is part of a large-scale effort to migrate google3's default C++ version from 17 to 20.

PiperOrigin-RevId: 500981197