Add "std::thread" compatible `mozc::Thread2`.
Mozc currently has `mozc::Thread` for multithreading, and it calls platform specific APIs internally (e.g. pthread in Linux).

It was created back in 2008, but today we have `std::thread` introduced in C++11, to which we can delegate handling platform differences.

This commit introduces `mozc::Thread2` (to be renamed to `mozc::Thread` after migration) that exposes a subset of `std::thread` APIs, which will be implemented using `std::thread` in a trivial way.

For now, we use the existing `mozc::Thread` as its underlying implementation to make sure that threading implementation is consistent in a process, so that users of `mozc::Thread` can be incrementally migrated to `mozc::Thread2`.

The `CreateThreadWithDoneNotification` is a wrapper factory function to support `mozc::Thread::IsRunning` use cases.

PiperOrigin-RevId: 504791582