Implement PlatformString and PlatformStringView
PlatformString is an alias to std::basic_string, and PlatformStringView is a null-terminated platform-native string view. These classes aim to address a few pain points:

- absl::string_view doesn't guarantee null-termination and cannot be used for C-style APIs.
- Sometimes multiple conversions happen when calling APIs on Windows.
- consolidate platform conditionals into one build rule

<path/to/mozc>/base/file_stream.cc has ToPlatformString and this is a more complete version of that implementation.

PlatformStringView implements most of the C++17 string_view member functions, but lacks some operations like find, find_first_of, etc because the focus of the class is integration with C APIs. We can add them in the future if deemed necessary.

PiperOrigin-RevId: 516336863