Delete `DisableLegacyMozcForCurrentUserOnWin8()`
This is a logical rollback of changelist 39183361.

#### Reason for rollback

`MigrationUtil::DisableLegacyMozcForCurrentUserOnWin8()` was added more than 10 years ago as a quick workaround for users who upgraded from Windows 7 or prior OSes.  Let's roll it back as we no longer support such upgrading scenario.

#### Original change description

Add a quick workaround to unregister IMM32 Mozc on Windows 8.

---

PiperOrigin-RevId: 516419611