Remove `SHOW_INFOLIST_IMMEDIATELY` workaround
This CL effectively reverts cl/27182540, which was added as a workaround for GNU Emacs 22 and Meadow 3.0 as mentioned in the comment around `SHOW_INFOLIST_IMMEDIATELY`.

> Some applications such as NTEmacs22 or Meadow3.0 automatically and
> frequently generates `WM_IME_CONTROL`/`IMC_SETCOMPOSITIONWINDOW` messages.
> In this case, the renderer might not be able to show the InfoList
> because new rendering messages will come before the InfoList is
> displayed with the default delay (~500 msec). This compatibility flag
> represents that we should show the InfoList immediately on such
> applications. See b/5824433 for details.

Now that IMM32 Mozc is gone, `WM_IME_CONTROL` and `IMC_SETCOMPOSITIONWINDOW` are something that CUAS is responsible for taking care of, and apps might have gotten updates, it's worth checking the current status again.

It turns out that Meadow website no longer exists, and GNU Emacs 28.1 running on Windows 11 with TIP Mozc doesn't require this workaround any more.  So let's roll that workaround back for simplicity.

#codeheahlth

PiperOrigin-RevId: 519938755