Stop using merge-module to redistribute VC runtime
This is a preparation to support building installers for Windows (#723).

Merge modules for Visual C++ redistributable files are deprecated in Visual Studio 2019 and later versions.

https://learn.microsoft.com/en-us/cpp/windows/redistributing-components-by-using-merge-modules

To make our WiX script be compatible with Visual Studio 2019 and later versions, we need to install Visual C++ Runtime Library without relying on merge-module file (`*.msm`).

With this CL, the following DLLs will be installed without relying on merge-module unless `--noqt` option is specified in `build_mozc.py gyp`.

 - `msvcp140.dll`
 - `vcruntime140.dll`

The new implementation should be compatible with Visual Studio 2015, 2017, 2019 and 2022.

One notable difference from the previous implementation is that the installer no longer tries to install debug version of Visual C++ Runtime, which is actually not allowed in the license.  Other than that, there should be no user observable behavior difference.

#codehealth

PiperOrigin-RevId: 525134892