Remove special condition to trigger GitHub Actions
This is a logical revert of 48f66f4666279f88ca6543ba26eb2132e4f74625, which was made for #729.

It turns out that for public repositories there is no cost / resource implications for running GitHub Actions.

To simplify the workflow rules, let's not have any special condition to trigger GitHub Actions for now.

PiperOrigin-RevId: 526597280