Build Qt from the source code by default for macOS build
Realistically, for macOS and Windows, providing our own scripts to build Qt from the source code is supposed to work better than asking contributors to pick up random versions of libraries with random build configurations in the build instructions.  With this CL, the way how to set up Qt build dependency is now fully managed with checked-in Python scripts.

Here are notable changes:

* Updated macOS build instructions.
  - Bazel build starts depending on Python 3.9+ again.  Hopefully this dependency can be removed in the future by downloading Qt source code and build it with Bazel build rules.
  - Simplified explanations about Qt with `update_deps.py` and `build_qt.py`.
* Updated macOS workflow for GitHub Actions.
  - Qt is now built as a build step.  Net increase of turnaround time is roughly 19 min.
  - Increased the build timeout from 60 min to 90 min. The total built time is still around 45 min or so.

Closes #722.

PiperOrigin-RevId: 528373593