Make `IPCTest` deterministic
This is a follow up for #741 and #747.

From the viewpoint of platform abstraction, it'd make more sense to first verify how much data can be reliably handled in our IPC infrastructure. In that sense, using predictable test dataset with much lower entropy is considered to be not a major concern in `IPCTest`.

With that, this CL simplifies `GenerateInputData` by mostly filling the data with 'x'.

#codehealth

PiperOrigin-RevId: 535503372