Fix warning messages in GitHub Actions for Win
This is a follow up to 1486014dc54aaf9d81df953e646facd635731f46 for #749.

Putting a line starts with `#` in `run` lines ends up with the following (non-critical) warning for `cmd` shell in Windows.  Let's move it outside of `run:` lines to keep the build log cleaner.

```
'#' is not recognized as an internal or external command,
operable program or batch file.
```

There should be no actual behavior change in the build process itself.

PiperOrigin-RevId: 536563466