Assume Qt is available when building the installer for Windows
This is a preparation to support code-signing and building installers on a different machine than build machines.

With this CL, building installers for Windows starts requiring Qt binaries.

While there are still several use cases for us to be able to build installers for Windows without Qt dependencies, supporting it have largely contributed to the complexity in our `*.wxs` files.

To make on-going migrations such as build system migration from Gyp to Bazel easier, let's reduce the number of supported build scenarios and simplify `*.wxs` files.

This is just an internal code clean up.  There must be no user observable behavior change.

#codehealth

PiperOrigin-RevId: 538391809