Use the actual DLL name for binary ID name
When the `dark.exe` in the WiX toolkit decompiles MSI files, it uses the ID name specified to the `<Binary>` element when extracting binary files.

For instance, the file content associated with
```
<Binary Id='CustomActions64' SourceFile='$(var.CustomActions64Path)' />
```
will be extracted into
```
./Binary/CustomActions64
```

By using the actual DLL name as the ID, our custom action DLLs will be extracted with the original file name, which might be helpful for certain type of operations such as code signing.

In the installation process, ID is just an ID and there must be no user observable behavior change.

PiperOrigin-RevId: 538967593