Add mozc_run_build_tool Bazel rule and embed_file macro, clean up build scripts
mozc_run_build_tool is a new Bazel rule to address the limitation of Skylib's run_binary, which doesn't support multiple files in a single label. It's designed to work with many of the Mozc's build scripts and works on Windows. It's not going to cover all use cases, but it works as a complement to run_binary.

embed_file macro runs embed_file.py. It's invoked often so having a simpler macro helps.

py scripts: use argparse and change to produce constexpr constants.

PiperOrigin-RevId: 541687622