Set SelectionShortcut.SHORTCUT_123456789 always for Gboard client
Previously this config was set only when a key event was from a PK (not worked correctly though; the cause of b/279402880 was that when SimpleJapaneseIme#getShortcutMode() was called, currentKeyboardSpecification
was not set yet so the shortcut was always disabled),
but shortcut behavior should always be no-op for VK events and be effective for PK events.

This CL always enable the shortcut config, and in the decoder
the data source of a key event is checked on conversion state.
If the current request has a special romaji table (== the key event is from VK),
the decoder treats it not as shortcut selection but character insertion.

Startblock:
   tap android_ime_all_platforms is green
PiperOrigin-RevId: 545906463