Remove unnecessary Qt configure options (#775)
It turns out that '-skip' option is supported only for the top-level
build thus does not work for submodules such as qtbase.  Moreover
having such '-skip' options results in a hard failure in Qt6.

Also '-nomake tools' is also turned out to have no effect.  Having
this results in a warning message for Qt6.

#codehealth

PiperOrigin-RevId: 550116309