Make sure to exclude session_watch_dog_test from mac build
This is a follow up to 6db68dbcc990db867f9aa4420f3c1adf049c7757.

As mentioned in the document [1], --build_tag_filters doesn't affect
the set of tests executed with the 'test' command.  We need to use
'--test_tag_filters' instead.

With this commit 'session_watch_dog_test', which is known to be flaky
in mac environment, will be excluded from mac CI as intended [2].

 [1]: https://bazel.build/reference/command-line-reference#flag--build_tag_filters
 [2]: be1dc58e35806803fb2de0b22f502ab80cb0fb76

PiperOrigin-RevId: 552043275