Replace `Qt5` with `Qt` in MSI component names
This is a preparation for Qt6 migration (#775).

In order to make *.wxs files compatible with both Qt5 and Qt6, it'd make more sense to not have Qt version '5' in MSI component names.  Let's simply drop '5' there, e.g. from 'Qt5CoreDll' to 'QtCoreDll'.

There must be no behavior change in this commit.

PiperOrigin-RevId: 552284979