Fully switch from Ubuntu 20.04 to Ubuntu 22.04
This follows up to 98060be03fb6a3ca865d261ea171bff43853f826.

With this commit we use Ubuntu 22.04 as the reference build environment
for Mozc for Linux and Android.

This commit also updates NDK version from 'r21b' to 'r25c' as otherwise
Android build fails.

Also the default Clang version in Ubuntu 22.04 is '14' thus we do not
need to explicitly set them in 'CC' and 'CXX'.

This is also a preparation of Qt6 migration (#775), as Ubuntu 20.04 does
not have Qt6 in their official repository.

Closes #763.

PiperOrigin-RevId: 552728127