Fix no candidate window in Windows
This is a follow up to da2630a882adcc1a85dce0c94318bd9dc6ccb3fa, which
accidentally introduced a deterministic access violation in
'mozc_renderer' in Windows.

The root cause is an unintentional behavior change in 'CreateFormat'.
This commit makes it clear that the buffer size needs to be carefully
specified.

Closes #788.

PiperOrigin-RevId: 554186217