Skip extracting most of Qt test files
As discussed in #769, storage limit is somewhat critical for GitHub
Action runners. As for Qt source code, we already skip extracting
several files that are not necessary to build Qt binaries used by Mozc.
With this commit, more unnecessary test files are skipped when
extracting Qt source code.

Here are net size changes of extracted source code.

  Qt 5.15.10: 262 MB -> 136 MB
  Qt 6.5.2  : 211 MB -> 119 MB

This is just an optimization of build process. There must be no user
observable behavior change.

PiperOrigin-RevId: 554704573