Fix #795: Add candidate window config to 'ibus_config.textproto'
This is a follow up to d40dc44d8a34cbe7946a3e06f37100775b2d517f,
which introduced the 'MOZC_IBUS_CANDIDATE_WINDOW' environment variable
per discussion at #637.

With this commit, 'ibus_config.textproto' can be used to customize
when to use 'mozc_renderer' instead of relying on the environment
variable.

To disable Mozc' candidate window:

  mozc_renderer {
    enabled : False
  }

To use Mozc' candidate window for X11 sessions and KDE Wayland
sessions:

  mozc_renderer {
    enabled : True
    compatible_wayland_desktop_names : ["KDE"]
  }

If there is no settings, it will be interpreted as if there was the
following settings.

  mozc_renderer {
    enabled : True
    compatible_wayland_desktop_names : ["GNOME"]
  }

While environment variable 'MOZC_IBUS_CANDIDATE_WINDOW' is still
supported, it is considered to be deprecated and may be removed in a
future version of Mozc.

PiperOrigin-RevId: 560186593