Make 'word_register_dialog.cc' compatible with 64-bit build
This is a preparation to start building 'mozc_tool' as a 64-bit binary
for Windows (#732).

The last parameter of 'SendMessageTimeout' is not 'DWORD*' but
'DWORD_PTR*', thus declaring 'message_response' as 'DWORD' results in a
compile error when building it for 64-bit.

That said, the last parameter is optional and we do not use the return
value anyway.  Let's just pass 'nullptr' there.

#codehealth

PiperOrigin-RevId: 561834120