#730: Also upload 'libmozcjni.so' as artifacts
This is a follow up to my commit [1], which enabled GitHub Actions
to upload build artifacts for Windows, macOS, and Linux but not for
Android.  This commit is to also support Android.

One interesting requirement for Android is that developers usually need
to build *.so for the following 4 architectures.

 - armeabi-v7a
 - arm64-v8a
 - x86
 - x86_64

With this commit, our GitHub Actions rule actually starts building
'libmozcjni.so' for the above 4 architectures, then archives them into
a single zip file and upload it as the final artifact.

 [1]: 2b04845f7b379607daa8315fb76bbc4ea76f1ca9

PiperOrigin-RevId: 561867498