Add '--macos_cpus' option to 'build_qt.py'
This is a preparation to support universal binaries for macOS (#801).

With this commit, you can start building Qt6 as universal binaries as follows.

 # Build universal binaries
 python3 build_tools/build_qt.py --release --confirm_license  \
                                 --macos_cpus=x86_64,arm64

Under the hood, '-DCMAKE_OSX_ARCHITECTURES' option is passed to the configure
options of Qt6 [1].

You can also build Intel64 binaries on arm64 Mac as follows.

 # Build Intel64 binaries
 python3 build_tools/build_qt.py --release --confirm_license  \
                                 --macos_cpus=x86_64

A known limitation is that building arm64 binaries on Intel64 mac does not
currently work though.

If '--macos_cpus' is not specified, Qt6 binaries will be built for the host CPU
architecture as it used to be.

 [1]: https://doc.qt.io/qt-6.5/macos-building.html#step-2-build-the-qt-library

PiperOrigin-RevId: 562439404