Simplify 'code_generator_util.py' a bit
This is a preparation before simplifying 'common.gypi' and 'common_win.gypi'
for #803.

As Visual C++ 2022 does not have a 65535-character limitation in string
literals, 'code_generator_util.py' does not need to change how to embed
data as a C++ byte stream depending on the target compiler.  Therefore
we no longer need to propagate the target compiler name in data_manager.gypi.

There must be no observable behavior change in the final artifacts.

#codehealth

PiperOrigin-RevId: 563309376