Remove '--warn_as_error' option from 'build_mozc.py'
As GYP build is basically deprecated except for Windows, there shouldn't be a
big problem in removing '--warn_as_error' option from 'build_mozc.py', which has
been off by default in OSS build.

This is about compiler warnings.  There must be difference in the build
artifacts.

This is a preparation for #803.

#codehealth

PiperOrigin-RevId: 563446029